requirements.txt
# Core dependencies
gradio>=4.19.2
langchain>=0.1.9
langchain-core>=0.1.23
langchain-openai>=0.0.5
langchain-community>=0.0.19
langchain-chroma>=0.0.1
langchain-google-genai>=0.0.6

# Database and data handling
pandas>=2.2.1
pyodbc>=5.0.1

# OpenAI and embeddings
openai>=1.12.0
tiktoken>=0.6.0

# Environment and utilities
python-dotenv>=1.0.0
chromadb>=0.4.22
numpy>=1.26.4

# Development tools
pytest>=8.0.2
black>=24.2.0
flake8>=7.0.0

# Additional dependencies
reportlab>=4.1.0
google-generativeai>=0.3.2 